Explanation of Weight Parameters Added
A short explanation of the concept of weight, which was missing so far, was added to the Concepts session. "Weights are values that control the strength of the connection between two neurons. That is, inputs are typically multiplied by weights, and that defines how much influence the input will have on the output. In other words: when the inputs are transmitted between neurons, the weights are applied to the inputs along with an additional value (the bias)".